+++ src/OpenTelemetry.Exporter.OpenTelemetryProtocol/OtlpTraceExporter.cs
@@ -1,4 +1,4 @@
-// <copyright file="OtlpExporter.cs" company="OpenTelemetry Authors">
+// <copyright file="OtlpTraceExporter.cs" company="OpenTelemetry Authors">
 // Copyright The OpenTelemetry Authors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -35,7 +35,7 @@ namespace OpenTelemetry.Exporter.OpenTelemetryProtocol
     /// Exporter consuming <see cref="Activity"/> and exporting the data using
     /// the OpenTelemetry protocol (OTLP).
     /// </summary>
-    public class OtlpExporter : BaseExporter<Activity>
+    public class OtlpTraceExporter : BaseExporter<Activity>
     {
         private const string DefaultServiceName = "OpenTelemetry Exporter";
 
@@ -49,20 +49,20 @@ namespace OpenTelemetry.Exporter.OpenTelemetryProtocol
         private readonly Metadata headers;
 
         /// <summary>
-        /// Initializes a new instance of the <see cref="OtlpExporter"/> class.
+        /// Initializes a new instance of the <see cref="OtlpTraceExporter"/> class.
         /// </summary>
         /// <param name="options">Configuration options for the exporter.</param>
-        public OtlpExporter(OtlpExporterOptions options)
+        public OtlpTraceExporter(OtlpExporterOptions options)
             : this(options, null)
         {
         }
 
         /// <summary>
-        /// Initializes a new instance of the <see cref="OtlpExporter"/> class.
+        /// Initializes a new instance of the <see cref="OtlpTraceExporter"/> class.
         /// </summary>
         /// <param name="options">Configuration options for the exporter.</param>
         /// <param name="traceServiceClient"><see cref="OtlpCollector.TraceService.TraceServiceClient"/>.</param>
-        internal OtlpExporter(OtlpExporterOptions options, OtlpCollector.TraceService.ITraceServiceClient traceServiceClient = null)
+        internal OtlpTraceExporter(OtlpExporterOptions options, OtlpCollector.TraceService.ITraceServiceClient traceServiceClient = null)
         {
             this.options = options ?? throw new ArgumentNullException(nameof(options));
             this.headers = options.Headers ?? throw new ArgumentException("Headers were not provided on options.", nameof(options));